kinda works now

+++ App/Inc/control.h
@@ -14,9 +14,9 @@ void Oven_start(void);
 void Oven_stop(void);
 PID Oven_get_PID(void);
 void Oven_set_PID(PID pid);
-uint32_t Oven_temperature(void);
-void Oven_control(uint32_t current_temperature);
-void Oven_set_target(uint32_t temperature);
+uint16_t Oven_temperature(void);
+void Oven_control(uint16_t current_temperature);
+void Oven_set_target(uint16_t temperature);
 int32_t Oven_error(void);
 
 #endif
